+++ kibot/misc.py
@@ -301,6 +301,8 @@ W_ERCJSON = '(W141) '
 W_ERC = '(W142) '
 W_DEPR = '(W143) '
 W_FILXRC = '(W144) '
+W_DRC = '(W145) '
+W_DRCJSON = '(W146) '
 # Somehow arbitrary, the colors are real, but can be different
 PCB_MAT_COLORS = {'fr1': "937042", 'fr2': "949d70", 'fr3': "adacb4", 'fr4': "332B16", 'fr5': "6cc290"}
 PCB_FINISH_COLORS = {'hal': "8b898c", 'hasl': "8b898c", 'imag': "8b898c", 'enig': "cfb96e", 'enepig': "cfb96e",